-- fix: non StrictCompile() fixes throwing errors on compile, mostly due to mis-defined callbacks

+++ src/libs/python/libbongo/json.h
@@ -38,7 +38,7 @@ void JsonObject_dealloc(JsonObject *self);
 PyObject *JsonObject_str(PyObject *selfp);
 
 /* Mapping protocol */
-int JsonObject_length(PyObject *self);
+Py_ssize_t JsonObject_length(PyObject *self);
 PyObject *JsonObject_subscript(PyObject *self, PyObject *key);
 int JsonObject_subscript_assign(PyObject *self, PyObject *key, PyObject *value);
 
@@ -59,16 +59,16 @@ extern PyTypeObject JsonArrayType;
 
 /* Sequence protocol */
 
-int JsonArray_length(PyObject *selfp);
+Py_ssize_t JsonArray_length(PyObject *selfp);
 PyObject *JsonArray_concat(PyObject *selfp, PyObject *b);
-PyObject *JsonArray_repeat(PyObject *selfp, int i);
-PyObject *JsonArray_item(PyObject *selfp, int i);
-PyObject *JsonArray_slice(PyObject *selfp, int i1, int i2);
-int JsonArray_assign(PyObject *selfp, int i, PyObject *obj);
-int JsonArray_assign_slice(PyObject *selfp, int i1, int i2, PyObject *obj);
+PyObject *JsonArray_repeat(PyObject *selfp, Py_ssize_t i);
+PyObject *JsonArray_item(PyObject *selfp, Py_ssize_t i);
+PyObject *JsonArray_slice(PyObject *selfp, Py_ssize_t i1, Py_ssize_t i2);
+int JsonArray_assign(PyObject *selfp, Py_ssize_t i, PyObject *obj);
+int JsonArray_assign_slice(PyObject *selfp, Py_ssize_t i1, Py_ssize_t i2, PyObject *obj);
 int JsonArray_contains(PyObject *selfp, PyObject *b);
 PyObject *JsonArray_inplace_concat(PyObject *selfp, PyObject *b);
-PyObject *JsonArray_inplace_repeat(PyObject *selfp, int i);
+PyObject *JsonArray_inplace_repeat(PyObject *selfp, Py_ssize_t i);
 
 
 PyObject *JsonObjectToPy(BongoJsonObject *object, BOOL own);
